HTET/requirements.txt
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PySimpleGUI==4.57.0
pytesseract==0.3.9
